MFC: r366229
For mulitcons boot, report it and which console is primary

Until we can do proper /etc/rc output on both consoles in multicons
boot (or all of them if we ever generalize), report when we are
booting multicons. Also report the primary console. This will be a big
hint why output stops after this line (though some slow USB discovery
still happens after mountroot / init starts).

Reviewed by: scottl@, tsoome@
Differential Revision: https://reviews.freebsd.org/D26574


git-svn-id: https://svn.freebsd.org/base/stable/12@366423 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f