Avoid using single quotes in arguments to logger.
Single quotes interfere with the workaround put in with r335753 and
aren't necessary in this case. I believe that all the underling issues
with r335753 have been corrected, but need to do more extensive
followup before reverting it as a bad idea.

PR: 240411
MFC After: 2 days (to give it time to get into 12.2)


git-svn-id: https://svn.freebsd.org/base/head@366578 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f