MFC r366647:
pf: create a kif for flags

If userspace tries to set flags (e.g. 'set skip on <ifspec>') and <ifspec>
doesn't exist we should create a kif so that we apply the flags when the
<ifspec> does turn up.

Otherwise we'd end up in surprising situations where the rules say the
interface should be skipped, but it's not until the rules get re-applied.


git-svn-id: https://svn.freebsd.org/base/stable/12@367056 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f