MFC r366667:
pf: do not remove kifs that are referenced by rules

Even if a kif doesn't have an ifp or if_group pointer we still can't delete it
if it's referenced by a rule. In other words: we must check rulerefs as well.

While we're here also teach pfi_kif_unref() not to remove kifs with flags.

Reported-by: syzbot+b31d1d7e12c5d4d42f28@syzkaller.appspotmail.com


git-svn-id: https://svn.freebsd.org/base/stable/12@367057 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f