MFC r366648:
pf tests: Test that 'set skip on <group>' works on new group members

There's a know issue where new group members don't get the 'set skip on'
applied until the rules are re-loaded.

Do this by setting rules that block all traffic, but skip members of the
'epair' group. If we can communicate over the epair interface we know the set
skip rule took effect, even if the rule was set before the interface was
created.


git-svn-id: https://svn.freebsd.org/base/stable/12@367058 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f